+++ tests/NumArray/GetNDimTest.php
@@ -1,0 +1,24 @@
+<?php
+declare(strict_types=1);
+
+namespace NumPHPTest\NumArray;
+
+use NumPHP\NumArray;
+
+class GetNDimTest extends \PHPUnit_Framework_TestCase
+{
+    public function testGetNDim4()
+    {
+        $numArray = new NumArray([8, 9, 3, 6]);
+        $this->assertSame(1, $numArray->getNDim());
+    }
+
+    public function testGetNDim2x3()
+    {
+        $numArray = new NumArray([
+            [8, 6, 4],
+            [3, 8, 0]
+        ]);
+        $this->assertSame(2, $numArray->getNDim());
+    }
+}