+++ tests/NumArray/OnesLikeTest.php
@@ -1,0 +1,27 @@
+<?php
+declare(strict_types=1);
+
+namespace NumPHPTest\NumArray;
+
+use NumPHP\NumArray;
+
+/**
+ * @SuppressWarnings(PHPMD.StaticAccess)
+ */
+class OnesLikeTest extends \PHPUnit_Framework_TestCase
+{
+    public function testOnesLike4()
+    {
+        $numArray = new NumArray([6, -6, -1, -7]);
+        $this->assertTrue(NumArray::ones(4)->isEqual(NumArray::onesLike($numArray)));
+    }
+
+    public function testOnesLike2x3()
+    {
+        $numArray = new NumArray([
+            [0, 5, -8],
+            [4, 8, 1]
+        ]);
+        $this->assertTrue(NumArray::ones(2, 3)->isEqual(NumArray::onesLike($numArray)));
+    }
+}